+++ packages/engine/Source/Shaders/Builtin/Constants/passCesium3DTilePlanarFillId.glsl
@@ -1,0 +1,9 @@
+/**
+ * The automatic GLSL constant for {@link Pass#CESIUM_3D_TILE_PLANAR_FILL_ID}
+ *
+ * @name czm_passCesium3DTilePlanarFillId
+ * @glslConstant
+ *
+ * @see czm_pass
+ */
+const float czm_passCesium3DTilePlanarFillId = 5.0;